Moving shape definition to a shared location
